requirements.txt
# SUPER FAST FREE RAG Service Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
requests>=2.31.0
pdfplumber>=0.9.0
nltk>=3.8.0
pinecone-client>=2.2.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
numpy>=1.21.0
